access-control: add acl api feature
move these commonly re-implemented api endpoints to this shared crate
so they can be easily re-used.

a user of the crate will need to extend the `AccessControlConfig` with
the following three functions:

- `acl_audit_privilege()`: returns the privilege necessary to see all
  acl entries
- `acl_modify_privilege()`: returns the privilege necessary to edit
  the acl tree beyond a user's api token privileges
- `check_acl_path()`: checks whether a path is a valid acl path in the
  context of the product using this crate. by default all paths are
  considered valid.

all three provide default implementations so that users that only use
the `impl` feature don't need to change anything.

Signed-off-by: Shannon Sterz <s.sterz@proxmox.com>
Link: https://lore.proxmox.com/20250828105912.294887-3-s.sterz@proxmox.com