fix PVE8 node status API call
PVE9 added a new non-optional field that does not exist in 8, so for now
overwrite that specific struct with a custom one where 'available' is
optional. This makes the deserialization work again for PVE8 too

In the future we need to think about how to handle different APIs
between version in a better way.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>