schema: use i64 for minimum / maximum / default integer values
IntegerSchema used isize for the minimum / maximum parameters. On 32
bit targets (wasm for instance) API defintions with minimum / maximum
sizes outside the i32 range would fail.

This also changes the u64 deserialize to fail if it encounters a value
that cannot be represented as i64 instead of simply casting it to i64
and moving on.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>