Schema2Rust: handle missing parameters->properties
We crashed if we encountered
```
    parameters => {
        additionalProperties => 0,
    },
```
almost all endpoint specify this explicitly, so
```
    parameters => {
        additionalProperties => 0,
        properties => {},
    },
```
I found at least these
- `/cluster/firewall/options`
- `/cluster/firewall/macros`
to be exceptions to this, there may be more.

Our api-macro does also not need this to be specified explicitly,
therefore we have the generator infer `properties => {}` in case it is
not specified explicitly, i.e. it is missing.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>