login: fix userid check for '@quarantine' user tickets
when logging into the pmg quarantine via LDAP, the user typically
enters a userid like 'foo@bar.com'. When receiving a valid ticket,
this contains a userid like 'foo@bar.com@quarantine'. To check if that's
correct, use our helper instead of manually checking for equality.

That helper also needs fixing: while it should be (optionally) possible
to enter the username with 'foo@bar.com@quarantine' (so we have to strip
the quarantine part from the expected userid), we also have to strip the
'@quarantine' part from the ticket response, since it contains that too.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Tested-by: Stoiko Ivanov <s.ivanov@proxmox.com>
Link: https://lore.proxmox.com/20251015142310.3633830-2-d.csapak@proxmox.com