s3-client: make storage class in list objects v2 response optional
Some providers (e.g. Google cloud) do not include the storage class
field in list objects v2 responses.

Make the storage field optional, so response parsing does not fail if
it is absent.

Since the PBS S3 backend does not use this field, and is currently
the only user of the crate, no further code adaptions are necessary.

Reported-by: https://forum.proxmox.com/threads/173256/
Signed-off-by: Christian Ebner <c.ebner@proxmox.com>
Link: https://lore.proxmox.com/all/20251020091559.106203-1-c.ebner@proxmox.com