pve-api-types: add regex for both storage- and bridge-pair
These are used for target storage and bridge mappings in remote
migrations. For the list items we don't really need a verifier function
so remove them and add regexes instead.

What this is not doing is check for uniqueness of mappings in the list,
but that is not really possible. Since we can't currently have verifiers
on fields that are not strings.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>