pve-api-types: rework -list to array handling
The implementation was very fragile and did regex-matching on types,
textually, without using any of the information available about the
type.
Instead of identifying string-lists as arrays in `type_of`, which is
not meant to be this magical, handle the case in `handle_def` and
pass the actual type-change information down. This means we can get
rid of several added regex matches, as the info is already there. The
codegen now handles optional vs non-optional as well (but untested,
since we don't have such cases for the code yet).

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>