api-macro: support an untagged "other" string-enum variant
This allows adding a variant of the form

    #[serde(untagged)]
    Other(String)

to a string enum with an `#[api]` macro. Note that doing this turns
off validation for the enum variants, since we cannot know whether a
variant was a typo or an actually unknown variant. The known variants
are still shown in the documentation as the (only) accepted ones.

Signed-off-by: Wolfgang Bumiller <w.bumiller@proxmox.com>