apt: update to old_version becoming an Option in newer apt-api-types
This adapts the proxmox-apt crate to recent changes in
proxmox-apt-api-types by commit ef17b5db ("apt-api-types: make
old_version optional"), which made the old_version field an Option to
correctly represent reality.

Single slightly questionable choice is using None for the unknown
package, but this can still be changed, and starting out with None
here seems not really wrong, especially as it might be better to
"catch" this at a higher level, like wrapping the whole update info
struct into an enum. Anyhow, lets keep this decision for when we
actually have a need for more specific behavior, I just wanted to
document here that it doesn't inherently has to be None.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>