proxmox-fixed-string: implement hash trait
Some enums in pbs-api-types implement Hash, so FixedString needs to
implement it too in order to be usable there.

While it should have been fine to use the whole struct here because
the buffer is zeroed on construction and the length is constant for
equal strings - still only use the actual bytes from the buffer that
contain the current string to avoid potential shenanigans.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251120145031.550340-4-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>