pbs-api-types: add optional fallback variants to enums in public API
In order to be able to provide forwards-compatibility in the Proxmox
Datacenter Manager, add fallback variants to all enums that are
exposed in the public API. This prevents breakage on PDM side whenever
a new enum variant is added in PBS.

The fallbacks are feature-guarded, so PBS can use the strict enums
without having to handle any fallback variants. This is unnecessary,
because pbs-api-types and proxmox-backup are updated in tandem, so
proxmox-backup should never run into the situation of having to handle
unknown variants. Only the datacenter manager will use this feature
flag for now, since it needs to potentially handle unknown variants.

Signed-off-by: Stefan Hanreich <s.hanreich@proxmox.com>
Link: https://lore.proxmox.com/20251120145031.550340-6-s.hanreich@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>