pve-api-types: schema2rust: don't mark URL parameters as optional
When a URL path parameter (e.g. .../rules/{pos}) shares a name with a schema
field that is marked as optional, the generator incorrectly wrapped the
path parameter in `Option<T>`.

Remove the logic that applies the schema's optionality to URL parameters,
ensuring they are always generated as required types.

Signed-off-by: Hannes Laimer <h.laimer@proxmox.com>
Link: https://lore.proxmox.com/20251128080733.43974-1-h.laimer@proxmox.com