pve-api-types: add cluster options api call
so we can get e.g. the datacenter notes.

The return schema for this is currently not defined and adding it in
pve-manager is rather hard.

The reason for this is that while we have defined the datacenter config
schema for the update call, the properties are all property-strings. In
the GET call however, we parse these and return a nested object.

We'd either have to copy the whole schema and make sure it does not
diverge, or find some more elegant way to modify the return schema to
include the parsed properties (for some properties this even happens
nested, e.g. a list in a property string gets split into an actual
array).

Since this is not feasible at the moment, the easiest way is to simply
use a Value for now.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Link: https://lore.proxmox.com/20251128101848.1580111-3-d.csapak@proxmox.com
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>