pve api types: percent encode string params in url
when we generate an url for the pve api from something like

"/some/{foo}/thing/{bar}"

then the url we generate will be wrapped in a `format!` and the
variables `foo` and `bar` will be inserted into the url.

Sometimes, these parameters can contain invalid characters for a url and
we have to percent encode them.

To do that, we convert the url to one with unnamed parameters and add
them with percent encoding.

The order of the url_params is the same as in the url, so we don't
have to to a lookup or similar, simply iterating over that is enough.

This fixes an issue with UPIDs that contain the character '>'.
(e.g. move disk from one vm to another).

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Reviewed-by: Lukas Wagner <l.wagner@proxmox.com>
Tested-by: Lukas Wagner <l.wagner@proxmox.com>
Link: https://lore.proxmox.com/20251128130124.2738745-1-d.csapak@proxmox.com