deb-version: initial commit for simple debian version comparison crate
This crate tries to implement the ordering rules of Debian versions as
documented in [deb-version].

It's main focus is to provide a somewhat convenient interface without
pulling in a ton of dependencies.

While there exists already a [debversion-rs] crate with a similar
goal, it's not ideal for our uses cases due to having a slightly
different focus, i.e. besides the simple version comparison, it also
wants to be able to do more debian packaging related requirements and
upstream watches and operating on some debian DBs that provide
interactions with their package repos. This means that it has more
code and dependencies, and while some of them are feature guarded,
others aren't. Biggest offender here is the chrono crate, which can
panic during usage and thus is not desired to be used as transitive
dependency in our stack if possible in any way.

This crate was created independently of [debversion-rs], but due to
the Debian version comparison being a rather specific and small
problem and me having lightly skimmed at the code of the former, it is
very likely that we'll share some basic principles. In any case I'd
like to thank the author from [debversion-rs] for their work,
especially the extensive unit test cases, this implementation here
would be definitively had been a bit worse if it wasn't for
[debversion-rs] existence.

For the actual implementation it's best to just check the code, the
real "meat" is just ~200 lines anyway.

[deb-version]: https://manpages.debian.org/trixie/dpkg-dev/deb-version.7.en.html
[debversion-rs]: https://github.com/jelmer/debversion-rs

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>