deb-version: add cmp_versions to allow &str comparison without allocation
While versions are relatively short and allocation is not _that_
expensive, it's still not nice if one has to frequently allocate when,
e.g., sorting a (big) list of package versions.

Add a new public cmp_versions function that takes two `&str` slices
and returns how they order to each other, reusing the same infra as
the Version struct as much as possible.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>