apt-api-types: add optional unknown fallback variant to APTRepositoryHandle
This type keeps giving and causes trouble in PDM when trying to parse
the repo state for older major releases, like the still supported PVE
8. Albeit, to be fair, most enums have this problem, but this here
shouldn't be an enum in the first place.

Anyhow, as some sort of stop-gap that is better than throwing our
hands up, lets also adopt here a fallback "UnknownEnumValue" variant
for the APTRepositoryHandle enum.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>